changed state from SearchBar to global, rearranged test as follow up
