+++ src/components/SearchBar/__tests__/tests.js
@@ -1,6 +1,6 @@
 import { shallowMount, mount, createLocalVue } from '@vue/test-utils';
 import Vuex from 'vuex';
-import searchbar from '../../../store/SearchBar/index.js';
+import store from '../../../store/index.js';
 import SearchBar from '@/components/SearchBar/SearchBar.vue';
 
 //create a local instance of our vue
@@ -9,19 +9,6 @@ const localVue = createLocalVue();
 localVue.use(Vuex);
 
 describe('User enters a searchword into the searchbar', () => {
-  let modules, store;
-
-  modules = {
-    searchbar: searchbar
-  };
-
-  beforeEach(() => {
-    //mock our store
-    store = new Vuex.Store({
-      modules
-    });
-  });
-
   test('Test so that everything renders correctly', () => {
     const wrapper = shallowMount(SearchBar, { store, localVue });
     expect(wrapper.element).toMatchSnapshot();
@@ -34,9 +21,8 @@ describe('User enters a searchword into the searchbar', () => {
     //Act
     input.setValue('');
     await input.trigger('keyup');
-    await localVue.nextTick();
     //Assert
-    expect(searchbar.state.filteredProducts.length).toBeGreaterThan(0);
+    expect(store.state.filteredProducts.length).toBeGreaterThan(0);
   });
 
   test('Test so that searchword returns items with matching letters', async () => {
@@ -46,9 +32,8 @@ describe('User enters a searchword into the searchbar', () => {
     //Act
     input.setValue('goldfish');
     await input.trigger('keyup');
-    await localVue.nextTick();
     //Assert
-    expect(searchbar.state.filteredProducts).toContain('goldfish');
+    expect(store.state.filteredProducts).toContain('goldfish');
   });
 
   test('Test so that searchword with capital letters becomes lowercase', async () => {
@@ -58,9 +43,8 @@ describe('User enters a searchword into the searchbar', () => {
     //Act
     input.setValue('GOLDFISH');
     await input.trigger('keyup');
-    await localVue.nextTick();
     //Assert
-    expect(searchbar.state.filteredProducts).toContain('goldfish');
+    expect(store.state.filteredProducts).toContain('goldfish');
   });
 
   test('Test so that action getByThisKeyword in vuex is called on keyup from component', async () => {
