+++ src/store/SearchBar/actions.js
@@ -1,9 +1,17 @@
 const actions = {
   getByThisKeyword(context, keyWord) {
-    context.commit('getByKeyword', keyWord);
+    const search = keyWord.toLowerCase();
+    if (!search.length) {
+      context.rootState.filteredProducts = context.rootState.allProducts;
+    } else {
+      const foundProducts = context.rootState.allProducts.filter(item =>
+        item.match(search)
+      );
+      context.rootState.filteredProducts = foundProducts;
+    }
   },
   displayThisEntireArray(context) {
-    context.commit('displayEntireArray');
+    context.rootState.filteredProducts = context.rootState.allProducts;
   }
 };
