+++ src/store/SearchBar/index.js
@@ -4,25 +4,7 @@ import actions from './actions.js';
 
 const state = {
   namespaced: true,
-  //TODO: placeholder array, remove this array when we have array with actual products!
-  state: {
-    allProducts: [
-      'cat',
-      'dog',
-      'parrot',
-      'goldfish',
-      'horse',
-      'elephant',
-      'ant',
-      'snake',
-      'crow',
-      'cow',
-      'eagle',
-      'rhinoceros',
-      'chimpanzee'
-    ],
-    filteredProducts: []
-  },
+  state: {},
   mutations: mutations,
   getters: getters,
   actions: actions
