+++ src/store/SearchBar/mutations.js
@@ -1,18 +1,3 @@
-const mutations = {
-  getByKeyword(state, keyWord) {
-    const search = keyWord.toLowerCase();
-    if (!search.length) {
-      state.filteredProducts = state.allProducts;
-    } else {
-      const foundProducts = state.allProducts.filter(item =>
-        item.match(search)
-      );
-      state.filteredProducts = foundProducts;
-    }
-  },
-  displayEntireArray(state) {
-    state.filteredProducts = state.allProducts;
-  }
-};
+const mutations = {};
 
 export default mutations;
